1.1.0 - breaking changes! d'oh. But awesomer bits.
Guess I shoulda started with 0.1.0, but I didn't think I'd hack on this
anymore. No worries, no one is using it but me right now.

The hacking of Bundler is simpler now, AND more flexible, because the
prior code put together a Resolver instance too soon to adequately
accommodate scenarios where we put a newer version of a gem into the
source index, but want to simulate the natural non-update Bundler
behavior of keeping what was previously locked.

This sets up the ability to experiment with how Bundler deals with
updates in an attempt to figure out how hard conservation updates would
be.