Support for `update` command w/ existing lockfile.
Found a place the faked source was being merged with the one read from
the lockfile, so now tests that want to simulate that case will work.

This also adds a spec that demonstrates how `bundle update foo` will
also grab an update to any gems `foo` depends on, even though we didn't
request that. This refers to the conservative idea that has been floated
around the 'net.

https://github.com/bundler/bundler/issues/2016
http://makandracards.com/makandra/13885-how-to-update-a-single-gem-conservatively